Split wastewater.rs (712 lines) into wastewater/ subdirectory
Split into three sub-modules for maintainability:
- types.rs (121 lines): constants, WastewaterState resource, helper functions
- systems.rs (182 lines): update_wastewater, wastewater_health_penalty systems, plugin
- tests.rs (431 lines): all unit tests

mod.rs re-exports all public items so external callers are unaffected.

Closes #1389
