+++ crates/simulation/src/wastewater/tests.rs
@@ -1,285 +1,16 @@
-//! Wastewater and sewage collection system (WATER-007).
-//!
-//! Buildings generate sewage at 80% of their water consumption. Sewage treatment
-//! plants provide treatment capacity. If total sewage exceeds capacity, overflow
-//! discharges as raw sewage into nearby water bodies (increasing water pollution).
-//! Uncollected sewage near residential areas applies a health penalty to citizens.
-
-use bevy::prelude::*;
-use serde::{Deserialize, Serialize};
-
-use crate::buildings::Building;
-use crate::citizen::{Citizen, CitizenDetails, HomeLocation};
-use crate::config::{GRID_HEIGHT, GRID_WIDTH};
-use crate::grid::{CellType, WorldGrid};
-use crate::utilities::{UtilitySource, UtilityType};
-use crate::water_demand::WaterDemand;
-use crate::water_pollution::WaterPollutionGrid;
-
-// =============================================================================
-// Constants
-// =============================================================================
-
-/// Fraction of water consumption that becomes sewage.
-const SEWAGE_FRACTION: f32 = 0.80;
-
-/// Treatment capacity per sewage plant in gallons per day.
-const TREATMENT_CAPACITY_PER_PLANT: f32 = 50_000.0;
-
-/// Pollution amount added to each water cell within discharge radius when overflow occurs.
-const DISCHARGE_POLLUTION_AMOUNT: u8 = 15;
-
-/// Radius (in grid cells) around a sewage plant within which buildings are considered serviced.
-const SEWAGE_SERVICE_RADIUS: i32 = 20;
-
-/// Radius (in grid cells) around the city center used to find water cells for discharge.
-const DISCHARGE_SEARCH_RADIUS: i32 = 30;
-
-/// Health penalty per slow tick for citizens living near uncollected sewage.
-const HEALTH_PENALTY_PER_TICK: f32 = 1.5;
-
-/// Happiness penalty per slow tick for citizens living near uncollected sewage.
-const HAPPINESS_PENALTY_PER_TICK: f32 = 3.0;
-
-/// Distance (in grid cells) from a building without sewage service that triggers health penalty.
-const UNCOLLECTED_PENALTY_RADIUS: i32 = 5;
-
-// =============================================================================
-// Wastewater state resource
-// =============================================================================
-
-/// City-wide wastewater and sewage tracking resource.
-#[derive(Resource, Default, Clone, Debug, Serialize, Deserialize)]
-pub struct WastewaterState {
-    /// Total sewage generated across all buildings in gallons per day.
-    pub total_sewage_generated: f32,
-    /// Total treatment capacity from all sewage plants in gallons per day.
-    pub total_treatment_capacity: f32,
-    /// Untreated sewage overflow amount in gallons per day (excess beyond capacity).
-    pub overflow_amount: f32,
-    /// Fraction of buildings with sewage service coverage (0.0..=1.0).
-    pub coverage_ratio: f32,
-    /// Number of raw sewage discharge events (incremented each period with overflow).
-    pub pollution_events: u32,
-    /// Whether a health penalty is currently active due to uncollected sewage.
-    pub health_penalty_active: bool,
-}
-
-// =============================================================================
-// Helper functions
-// =============================================================================
-
-/// Compute sewage generation for a building based on its water demand.
-/// Returns sewage in gallons per day (80% of water consumption).
-fn sewage_for_demand(demand: &WaterDemand) -> f32 {
-    demand.demand_gpd * SEWAGE_FRACTION
-}
-
-/// Check if a building at (bx, by) is within the service radius of any sewage plant.
-fn is_serviced_by_sewage_plant(bx: usize, by: usize, sewage_plants: &[(usize, usize)]) -> bool {
-    for &(px, py) in sewage_plants {
-        let dx = (bx as i32 - px as i32).abs();
-        let dy = (by as i32 - py as i32).abs();
-        if dx + dy <= SEWAGE_SERVICE_RADIUS {
-            return true;
-        }
-    }
-    false
-}
-
-/// Find water cells near a set of sewage plant locations for discharge.
-/// Returns a list of (x, y) water cell coordinates.
-fn find_discharge_water_cells(
-    grid: &WorldGrid,
-    sewage_plants: &[(usize, usize)],
-) -> Vec<(usize, usize)> {
-    let mut water_cells = Vec::new();
-
-    // If there are sewage plants, search around them for water cells
-    // Otherwise, search from the center of the map
-    let search_centers: Vec<(usize, usize)> = if sewage_plants.is_empty() {
-        vec![(GRID_WIDTH / 2, GRID_HEIGHT / 2)]
-    } else {
-        sewage_plants.to_vec()
-    };
-
-    for &(cx, cy) in &search_centers {
-        for dy in -DISCHARGE_SEARCH_RADIUS..=DISCHARGE_SEARCH_RADIUS {
-            for dx in -DISCHARGE_SEARCH_RADIUS..=DISCHARGE_SEARCH_RADIUS {
-                let nx = cx as i32 + dx;
-                let ny = cy as i32 + dy;
-                if nx < 0 || ny < 0 || (nx as usize) >= GRID_WIDTH || (ny as usize) >= GRID_HEIGHT {
-                    continue;
-                }
-                let ux = nx as usize;
-                let uy = ny as usize;
-                if grid.get(ux, uy).cell_type == CellType::Water {
-                    water_cells.push((ux, uy));
-                }
-            }
-        }
-    }
-
-    water_cells.dedup();
-    water_cells
-}
-
-// =============================================================================
-// System
-// =============================================================================
-
-/// Main wastewater update system. Runs every slow tick.
-///
-/// - Queries buildings with `WaterDemand` to compute sewage generation (80% of water use)
-/// - Queries `UtilitySource` for `SewagePlant` to get treatment capacity and plant locations
-/// - Computes coverage ratio (fraction of buildings within service radius of a plant)
-/// - If overflow (sewage > capacity), discharges pollution to nearby water cells
-/// - Sets health penalty flag when residential buildings lack sewage service
-#[allow(clippy::too_many_arguments)]
-pub fn update_wastewater(
-    slow_timer: Res<crate::SlowTickTimer>,
-    mut wastewater: ResMut<WastewaterState>,
-    grid: Res<WorldGrid>,
-    mut water_pollution: ResMut<WaterPollutionGrid>,
-    buildings: Query<(&Building, &WaterDemand)>,
-    utilities: Query<&UtilitySource>,
-) {
-    if !slow_timer.should_run() {
-        return;
-    }
-
-    // --- Phase 1: Collect sewage plant positions and compute treatment capacity ---
-    let mut sewage_plants: Vec<(usize, usize)> = Vec::new();
-    let mut total_capacity: f32 = 0.0;
-
-    for utility in &utilities {
-        if utility.utility_type == UtilityType::SewagePlant {
-            sewage_plants.push((utility.grid_x, utility.grid_y));
-            total_capacity += TREATMENT_CAPACITY_PER_PLANT;
-        }
-    }
-
-    // --- Phase 2: Compute total sewage generation and coverage ---
-    let mut total_sewage: f32 = 0.0;
-    let mut buildings_total: u32 = 0;
-    let mut buildings_serviced: u32 = 0;
-    let mut residential_unserviced = false;
-
-    for (building, demand) in &buildings {
-        let sewage = sewage_for_demand(demand);
-        total_sewage += sewage;
-        buildings_total += 1;
-
-        let serviced =
-            is_serviced_by_sewage_plant(building.grid_x, building.grid_y, &sewage_plants);
-        if serviced {
-            buildings_serviced += 1;
-        } else if building.zone_type.is_residential() {
-            residential_unserviced = true;
-        }
-    }
-
-    let coverage_ratio = if buildings_total > 0 {
-        buildings_serviced as f32 / buildings_total as f32
-    } else {
-        1.0 // No buildings means full coverage (nothing to service)
-    };
-
-    // --- Phase 3: Compute overflow ---
-    let overflow = (total_sewage - total_capacity).max(0.0);
-
-    // --- Phase 4: If overflow, discharge pollution to nearby water cells ---
-    if overflow > 0.0 {
-        wastewater.pollution_events += 1;
-
-        let water_cells = find_discharge_water_cells(&grid, &sewage_plants);
-        // Scale pollution by overflow severity (more overflow = more pollution per cell)
-        let severity_mult = (overflow / TREATMENT_CAPACITY_PER_PLANT).clamp(0.5, 3.0);
-        let pollution_amount = (DISCHARGE_POLLUTION_AMOUNT as f32 * severity_mult) as u8;
-
-        for &(wx, wy) in &water_cells {
-            let idx = wy * water_pollution.width + wx;
-            water_pollution.levels[idx] =
-                water_pollution.levels[idx].saturating_add(pollution_amount);
-        }
-    }
-
-    // --- Phase 5: Update state ---
-    wastewater.total_sewage_generated = total_sewage;
-    wastewater.total_treatment_capacity = total_capacity;
-    wastewater.overflow_amount = overflow;
-    wastewater.coverage_ratio = coverage_ratio;
-    wastewater.health_penalty_active = residential_unserviced && buildings_total > 0;
-}
-
-/// Health and happiness penalty for citizens living near areas without sewage service.
-/// Citizens whose homes are in residential buildings not covered by a sewage plant
-/// suffer reduced health and happiness.
-pub fn wastewater_health_penalty(
-    slow_timer: Res<crate::SlowTickTimer>,
-    wastewater: Res<WastewaterState>,
-    mut citizens: Query<(&mut CitizenDetails, &HomeLocation), With<Citizen>>,
-    buildings: Query<&Building>,
-    utilities: Query<&UtilitySource>,
-) {
-    if !slow_timer.should_run() {
-        return;
-    }
-
-    // Only apply penalty if the system has detected uncollected sewage
-    if !wastewater.health_penalty_active {
-        return;
-    }
-
-    // Collect sewage plant positions
-    let sewage_plants: Vec<(usize, usize)> = utilities
-        .iter()
-        .filter(|u| u.utility_type == UtilityType::SewagePlant)
-        .map(|u| (u.grid_x, u.grid_y))
-        .collect();
-
-    // Build a set of residential building positions without sewage service
-    let unserviced_residential: Vec<(usize, usize)> = buildings
-        .iter()
-        .filter(|b| b.zone_type.is_residential())
-        .filter(|b| !is_serviced_by_sewage_plant(b.grid_x, b.grid_y, &sewage_plants))
-        .map(|b| (b.grid_x, b.grid_y))
-        .collect();
-
-    if unserviced_residential.is_empty() {
-        return;
-    }
-
-    for (mut details, home) in &mut citizens {
-        let hx = home.grid_x;
-        let hy = home.grid_y;
-
-        if hx >= GRID_WIDTH || hy >= GRID_HEIGHT {
-            continue;
-        }
-
-        // Check if citizen's home is near any unserviced residential building
-        let near_unserviced = unserviced_residential.iter().any(|&(bx, by)| {
-            let dx = (hx as i32 - bx as i32).abs();
-            let dy = (hy as i32 - by as i32).abs();
-            dx + dy <= UNCOLLECTED_PENALTY_RADIUS
-        });
-
-        if near_unserviced {
-            details.health = (details.health - HEALTH_PENALTY_PER_TICK).max(0.0);
-            details.happiness = (details.happiness - HAPPINESS_PENALTY_PER_TICK).max(0.0);
-        }
-    }
-}
-
-// =============================================================================
-// Tests
-// =============================================================================
+//! Unit tests for the wastewater system.
 
 #[cfg(test)]
 mod tests {
-    use super::*;
-    use crate::grid::ZoneType;
+    use crate::config::{GRID_HEIGHT, GRID_WIDTH};
+    use crate::grid::{CellType, WorldGrid, ZoneType};
+    use crate::water_demand::WaterDemand;
+
+    use super::super::types::{
+        find_discharge_water_cells, is_serviced_by_sewage_plant, sewage_for_demand,
+        WastewaterState, DISCHARGE_POLLUTION_AMOUNT, HAPPINESS_PENALTY_PER_TICK,
+        HEALTH_PENALTY_PER_TICK, SEWAGE_FRACTION, TREATMENT_CAPACITY_PER_PLANT,
+    };
 
     // -------------------------------------------------------------------------
     // WastewaterState default tests
@@ -695,18 +426,5 @@ mod tests {
         assert!((deserialized.coverage_ratio - 0.75).abs() < 0.01);
         assert_eq!(deserialized.pollution_events, 5);
         assert!(deserialized.health_penalty_active);
-    }
-}
-
-pub struct WastewaterPlugin;
-
-impl Plugin for WastewaterPlugin {
-    fn build(&self, app: &mut App) {
-        app.init_resource::<WastewaterState>().add_systems(
-            FixedUpdate,
-            (update_wastewater, wastewater_health_penalty)
-                .after(crate::imports_exports::process_trade)
-                .in_set(crate::SimulationSet::Simulation),
-        );
     }
 }
